backend/app/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
pydantic[email]==2.4.2

# Database & ORM
supabase==2.1.2
sqlalchemy==2.0.23

# Data Processing
pandas==2.1.2
numpy==1.26.2
openpyxl==3.1.2

# API & HTTP
requests==2.31.0
httpx==0.25.1

# Payment Processing
razorpay==1.3.0

# Reporting & Export
reportlab==4.0.7
plotly==5.18.0

# Data Science & Analysis
scipy==1.11.4

# Utilities
python-slugify==8.0.1
pytz==2023.3

# Async
aiofiles==23.2.1

# Logging & Monitoring
python-json-logger==2.0.7

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Development (optional)
black==23.12.0
flake8==6.1.0
mypy==1.7.1
